requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.10
pytest==7.4.2
pytest-cov==4.1.0
coverage==7.3.2
requests
